requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
ultralytics==8.0.196
python-dotenv==1.0.0
pytest==7.4.2
requests==2.31.0
piexif==1.1.3
geopy==2.4.0
gunicorn==21.2.0
PyYAML==6.0.1